Only clamp frozen header rows/columns
Range clamping at the start of the range when unbounded is now not assuming a
fixed number of exactly one header row/column, and instead relying on the
number of *frozen* rows/columns (those that don't scroll with the rest of the
document).

I also added some documentation and comments in case I forget again, and
rewrote some of the code in a more modern style.