Simplify function to make it more robust and less surprising
The fallback to counting from the function's input argument if an error occured
while processing the formula made it difficult to be sure that it was computing
what was expected. By only supporting the formula-based use-case, it always
either does what is expected or fails visibly, giving the user the opportunity
to identify the issue and fix it.

The function doesn't support counting from its argument directly, but it's not
something I use, and not something I expect to be commonly needed.

The function now also indicates the faulty range ref in case of error resolving
it.