Use display value to ensure string type
Setting the data type of a cell to plain text is apparently not always
sufficient to actually get a string in the script with `range.getValues()`.

Switching data types nilly-willy in the spreadsheet can also lead to data loss
(try switching the type from automatic to plain text back and forth a few
times, with the initial value "1-2").

Retrieving the display value ensures that the value is always a string, but it
can be formatted based on the user's locale so can potentially cause problems
too with the parsing, e.g. a volume named "1.5" should get counted as 1 volume,
but if the locale uses a decimal comma instead of a decimal point it could be
displayed as "1,5" which is counted as 2 volumes. At least it's something the
user can see and therefore address.